backend/requirements.txt
# CertiTrack Backend Dependencies
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# PDF Generation
reportlab==4.0.8
weasyprint==60.2
Pillow==10.2.0

# QR Code
qrcode[pil]==7.4.2
python-barcode==0.15.1

# Task Queue & Scheduling
celery==5.3.6
redis==5.0.1
celery-redbeat==2.2.0

# Email & Notifications
aiosmtplib==3.0.1
jinja2==3.1.3

# Utils
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2

# ERPNext Integration (future)
frappe-client==2.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1

